Gets MD5 hashes of file in container and same file in bucket
This program needs to be able to check if files have changed so that they are not copied unnecessarily. This commit prints the hash for the same file when it is in the container's default storage and then when it is moved to the bucket. Note that once in the bucket the file is a google cloud 'blob'